Tended to @vidvidvid feedback
